+++ service/master/connection_from_service.h
@@ -13,7 +13,7 @@ public:
 
 	virtual void		onConnect(const std::string& szContext);
 	virtual void		onDisconnect();
-	virtual void		onDispatch(uint16_t nMsgType, const void* pData, uint16_t nSize);
+	virtual void		onDispatch(uint32_t nMsgType, const void* pData, uint16_t nSize);
 
 	const std::string&	getServiceName() const;
